librados: Added new API methods to create an ioctx by pool id
A new pool_list method includes the unique pool id for all pools.
This pool id can be used with the new rados_ioctx_create2 /
ioctx_create methods to create an ioctx by pool id instead of
by pool name.  Creating ioctx's by pool id helps avoid certain
race conditions when pools are renamed or deleted.

+++ src/librados/IoCtxImpl.h
@@ -52,7 +52,7 @@ struct librados::IoCtxImpl {
 
   IoCtxImpl();
   IoCtxImpl(RadosClient *c, Objecter *objecter, Mutex *client_lock,
-	    int poolid, const char *pool_name, snapid_t s);
+	    int64_t poolid, const char *pool_name, snapid_t s);
 
   void dup(const IoCtxImpl& rhs) {
     // Copy everything except the ref count
